test: update the data server test
