+++ tests/jax/integration/data_server_test.py
@@ -12,29 +12,20 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
 """Tests for data server for Jax-based Mava systems"""
 
-import functools
-from datetime import datetime
-from typing import Any
-
 import jax.numpy as jnp
-import launchpad as lp
-import optax
 import pytest
 import reverb
 from tensorflow.python.data.ops import dataset_ops
-
-from mava.systems.jax import System, mappo
-from mava.utils.environments import debugging_utils
-from mava.utils.loggers import logger_utils
+from mava.systems.jax import System
+from tests.jax.integration.mock_systems import mock_system_single_process
 
 
 @pytest.fixture
 def test_system() -> System:
-    """Mappo system"""
-    return mappo.MAPPOSystem()
+    """A built IPPO system"""
+    return mock_system_single_process()
 
 
 def get_dataset(data_server: reverb.client.Client) -> dataset_ops.DatasetV1Adapter:
@@ -50,66 +41,6 @@ def get_dataset(data_server: reverb.client.Client) -> dataset_ops.DatasetV1Adapt
 
 def test_data_server(test_system: System) -> None:
     """Test if the data server instantiates processes as expected."""
-
-    # Environment.
-    environment_factory = functools.partial(
-        debugging_utils.make_environment,
-        env_name="simple_spread",
-        action_space="discrete",
-    )
-
-    # Networks.
-    def network_factory(*args: Any, **kwargs: Any) -> Any:
-        return mappo.make_default_networks(  # type: ignore
-            policy_layer_sizes=(32, 32),
-            critic_layer_sizes=(64, 64),
-            *args,
-            **kwargs,
-        )
-
-    # Checkpointer appends "Checkpoints" to checkpoint_dir.
-    base_dir = "~/mava"
-    mava_id = str(datetime.now())
-    checkpoint_subpath = f"{base_dir}/{mava_id}"
-    # Log every [log_every] seconds.
-    log_every = 1
-    logger_factory = functools.partial(
-        logger_utils.make_logger,
-        directory=base_dir,
-        to_terminal=True,
-        to_tensorboard=True,
-        time_stamp=mava_id,
-        time_delta=log_every,
-    )
-    # Optimizer.
-    optimizer = optax.chain(
-        optax.clip_by_global_norm(40.0),
-        optax.adam(1e-4),
-    )
-
-    # Build the test_system
-    test_system.build(
-        environment_factory=environment_factory,
-        network_factory=network_factory,
-        logger_factory=logger_factory,
-        experiment_path=checkpoint_subpath,
-        optimizer=optimizer,
-        executor_parameter_update_period=10,
-        multi_process=False,
-        run_evaluator=True,
-        num_executors=1,
-        use_next_extras=False,
-        max_queue_size=5000,
-        sample_batch_size=2,
-        max_in_flight_samples_per_worker=4,
-        num_workers_per_iterator=-1,
-        rate_limiter_timeout_ms=-1,
-        nodes_on_gpu=[],
-        lp_launch_type=lp.LaunchType.TEST_MULTI_THREADING,
-        sequence_length=4,
-        period=4,
-    )
-
     (
         data_server,
         parameter_server,
